Hospital_System_Project/requirements.txt


PyQt5>=5.15.10
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1